% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pablo.R
\name{filtrar_df_col}
\alias{filtrar_df_col}
\title{Data.Frame Filters}
\usage{
filtrar_df_col(df, nombre_columna, eliminar_NA = TRUE, filtro = "todos", x = 2)
}
\arguments{
\item{df}{Un objeto de la clase data.frame}

\item{nombre_columna}{nombre de la columna que se desea filtrar}

\item{eliminar_NA}{Parametro que controla la eliminacion o no de NA, existe la posibilidad de modificar si son NULL, -9999.99, etc.}

\item{filtro}{tipo de filtro que queremos aplicar. Existiendo las siguientes posibilidades: mayor_media, menor_media, mayor_mediana, menor_mediana, rango_normal y atipicos}

\item{x}{establecimiento del valor numerico por el que se multiplica a las deviacion tipica para el calculo de valores atipicos}
}
\value{
lista con los df solicitados en filtro
}
\description{
\code {filtrar_df_col} filtrado por media, mediana y desviacion tipica de columnas de un df. Obteniendo listas con los diferentes df solicitados.
}
\examples{
\dontrun{
library(dplyr)
mi.df <- data.frame(ID = 1:7, Valor = c(10, 12, NA, 15, 100, 9, NA))
solo_atipicos <- filtrar_df_col(mi.df, "Valor", filtro = "atipicos")
}
}
