% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pablo.R
\name{filtrar_df_col}
\alias{filtrar_df_col}
\title{Data.Frame Filters}
\usage{
filtrar_df_col(df, nombre_columna, filtro = "todos", eliminar_NA = TRUE, x = 2)
}
\arguments{
\item{df}{Un objeto de la clase \code{data.frame}}

\item{nombre_columna}{Nombre de la columna que se desea filtrar}

\item{filtro}{Tipo de filtro que queremos aplicar. Opciones: "mayor_media",
"menor_media", "mayor_mediana", "menor_mediana", "rango_normal", "atipicos" o "todos".}

\item{eliminar_NA}{Booleano que controla la eliminacion o no de NA (na.rm)}

\item{x}{Valor numérico multiplicador para el cálculo de atípicos (Media +/- x*SD)}
}
\value{
Un data.frame filtrado o una lista si el filtro es "todos".
}
\description{
\code{filtrar_df_col} permite filtrar un dataframe basado en estadísticas descriptivas
(media, mediana, desviación típica) de una columna específica.
}
\examples{
\dontrun{
mi.df <- data.frame(ID = 1:7, Valor = c(10, 12, NA, 15, 100, 9, NA))
solo_atipicos <- filtrar_df_col(mi.df, "Valor", filtro = "atipicos")
}

}
